requirements.txt
amqp==5.0.6
anyio==3.1.0
anyjson==0.3.3
appdirs==1.4.4
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.6.4.0
bleach==3.3.0
celery==5.0.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.4
django-filter==2.4.0
django-hosts==4.0
django-multitenant==2.3.2
django-tenant-schemas==1.10.0
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
future==0.18.2
html5lib==1.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.2
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mysql-connector-python==8.0.25
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.2
ordered-set==4.0.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywin32==300
pywinpty==1.1.1
pyzmq==22.0.3
redis==3.5.3
reportlab==3.5.67
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
Send2Trash==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
xhtml2pdf==0.2.5
zope.interface==5.4.0
